OPEN AI 과금을 피하면서 스케줄러를 트리거하기 위해 수정 - 부하테스트 시나리오 중에 open ai 를 호출하는 부분이 있습니다. 과금을 피하기 위해 프로파일 설정, 인터페이스 생성을 했습니다. 또한 Jmeter에서 DailySummaryScheduler을 트리거할 수 없으니 스케줄러를 트리거하기 위해 TestSchedulerController를 만들어서 수동으로 엔드포인트를 노출시켰습니다.
