+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   profiles:
-    active: dev
+    active: dev, mock-openai
   messages:
     encoding: UTF-8
